[bounty] Pillory bounty redemption (#2658)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Attempt 2 at pillory bounty redemption, this time with a new sprite.

![image](https://github.com/user-attachments/assets/37a27ef2-f5d6-4d8b-94a6-092f4c474faa)
Replaces the town square, keep dungeon, and mercenary pillories with
excidium variants that let you slowly redeem bounties from criminals, 50
per 5 minutes. There is no way to speed this up, and you can't use this
to redeem the hunted quirk's bounty, though a case wasn't made for the
special. Tell me if I need to.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
A bounty that was requested for non-lethal redemption
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/29719e4f-e742-4f82-9799-79451587f2e8)

![image](https://github.com/user-attachments/assets/3ce02e52-6846-4daf-8fda-481da8ef02f3)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->