Fixes Permanently Slowed Simplemobs (#2868)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

- Fixes a bug where simplemobs would be permanently slow when getting
below a certain amount of health.

## Why It's Good For The Game

- Saigas do not need to be put down or retired anymore.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/f99d9628-8355-4504-83b9-50518777827c


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->