helps set characters voice type from before it was a value (#2869)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This should help those who have saved characters from before voice_type
was a value and auto set it based on sex if loading them brings no
value.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It might save admins the effort of changing it from ahelps and keep it
consistent for characters who probably should sound one way.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/001cb65f-5c1f-4ce3-87a1-85e38d5e1241


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->