Fix for necromancer-summoned skeletons preserving Swift intent (#2912)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR extends #2906 by removing the ability to preserve Swift intent
from skeletons raised with Raise Lesser Undead, which was missed from
the previous PR due to a silly oversight on my part.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
See #2906 

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/f09a2aa0-7a1a-4747-8ef3-c85ac8370a15)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->