Added logging to communication using vampyre telepathy (#2914)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR adds logging to vampyre telepathy. Previously, messages sent via
telepathy weren't visible in any of the round's log files; this PR logs
those communications in the Game log file.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Allows admins to better investigate issues which involve telepathic
communication between vampyres.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Telapathic message, as seen in-game:

![image](https://github.com/user-attachments/assets/d1731fda-79c4-47d6-ac9f-b5f8e952472e)

The same message, as seen in the log file:

![image](https://github.com/user-attachments/assets/015649b8-3393-41f0-80d8-9fed043608fa)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->