Added a record of sent Messages to the sending player's chat log (#2915)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR adds a record in the sender's chatlog of the contents of a
Message, when sent using the Message spell. It slightly modifies the
existing message on use of the spell to include the send message, with
the same font colour seen by the recipient.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
A simple quality-of-life feature - players can now see what Messages
they've sent, allowing them to recall exactly what was said and
potentially re-send it if they suspect it may not have correctly
appeared in the target's chat log.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/aa79b18e-37df-4d51-9094-a90dc4829f84)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->