Fixes cheesebun recipe (#2930)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Cheesebun recipe incorrectly needed an entire dough to be made, now
requires smalldough as it should
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Bug fix!
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/bf0586b6-6ecc-4865-bf11-1f14c73623a2)
![image](https://github.com/user-attachments/assets/1613f1b4-5adc-4efc-83e7-e6c917b1c406)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->