Allows for auto mining rock walls. (#2925)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Uses do_after to help a miner auto dig a tile based on toolspeed (I
think this is the same as the cooldown for its attack speed?)
I'm not entirely sure how to get the do_after to disable once you are
stunned from being out of stamina but it only tries to fire but fails.
It causes a double panting emote which isn't the worst thing in the
world but if anyone has an idea on how to fix that I could do it.

You still manually hit rocks, if you think I should add it to them I
could.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
my fingers are sore...
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/6427216b-816a-403a-a5cb-df8e113e7530


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->