removes extra frame from anklet (#2939)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
see title
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
it was causing flashing
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
bug shown

https://github.com/user-attachments/assets/cb0ccce8-bb29-4209-b8ac-6ade887d7563

fixed:


https://github.com/user-attachments/assets/23640cdc-795f-463b-8bc6-bbd449c072df



<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->