Fix for Death Knights who spawn via prompt not being properly initialised (#2933)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR fixes a longstanding issue with Death Knight spawning -
specifically, that Death Knights who spawn via the prompt, rather than
through the latejoin interface, spawn in a bugged state. These Death
knights would have no spells or skills, a generic name, and would be
alive rather than undead.

This issue was caused by a bad call to equipOutfit, which was passed, as
an argument, the Death Knight job, rather than the outfit. The proper
approach is to call SSJob.EquipRank, passing in the Death Knight, the
_title_ of the job, and an indication that the mob is latejoining. (This
boolean is essential to prevent a runtime error from attempting to
access new player information which is not guaranteed to exist.)

The forced movement of the mob to the vampyre lord has been moved after
the EquipRank call, to prevent the latejoining Death Knight from
appearing at one of the standard latejoin locations.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Prior to this PR, Death Knights spawning from the prompt would appear in
a bugged state, with no class stats, spells, skills, or traits, and
would spawn living rather than dead. This meant that less scrupulous
vampyre lords might turn them into spawn; more scrupulous ones might
instead put them toward manual labour or simply kill them, neither of
which are very interesting for the person who opts to become one. This
was not an issue with Death Knights who joined via the latejoin
interface, which is possible if no players responded to the prompt.

With the change, Death Knights who join via the prompt should function
correctly, rather than providing either an unlimited pool from which to
drink and create spawn. Death Knights are still limited in number to
three - a limit which may actually be reached now that they will
reliably add themselves to the list of Death Knights, since bugged Death
Knights did not. It's unclear to me whether this fix will prove to be a
buff to vampyres or not; I suppose time will tell, but after this
change, as with many of my fixes that touch elements of balance, we will
at least be able to tweak Death Knights with the knowledge that they
spawn as intended.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
A screenshot from a friend who aided me in testing, having spawned as a
Death Knight via the prompt. Note the presence of spells, Death Knight
equipment, undead eyes, lack of pain, and skeletonisation. Previously,
this would have simply been a human with some silly name like "Justice
Beach" (yes, this is an actual name that appeared as I was figuring out
this issue).

![image](https://github.com/user-attachments/assets/3d9d0762-b76e-45e1-a1dd-e390a4d2715f)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->