Added missing commas in german/grenzel replacements (#2942)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR fixes some syntax issues in the german_replacements.json file,
which were causing the Grenzel accent not to function correctly and
causing numerous runtime errors in the runtime logs.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Simple accent fix for players that choose to use the accent, and
eliminates a lot of runtime errors which were polluting the runtime
logs.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
My character saying "Goddess", "Hello", and "Goblins", respectively.

![image](https://github.com/user-attachments/assets/34572f74-09f4-4eee-aa36-ad573f186ad3)

![image](https://github.com/user-attachments/assets/5bb40978-7f7a-40c9-9ced-71380d4f55f8)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->