Map change fix (QOL) (#2874)
## About The Pull Request

Fixes map change not asking for a json file, resulting in maps not
rendering z levels properly.

## Why It's Good For The Game

Allows for seamless map changing without needing to make bothersome PRs
where you have to actually overwrite the rockhill.dm map file with the
new map. This is part of the BYOS event QOL fixes.

This is a general band-aid fix to this issue and it should work as
intended.

## Proof of Testing (Required)


![image_2025-05-17_201229348](https://github.com/user-attachments/assets/b45346dc-f831-43b1-aa45-21ef21305225)