Jam goblin encampment revamp (#2888)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
I rearranged the Goblin Encampment with a similar eye towards
player-traffic. Instead of a long, thin snaking corridoor, all things
connect to a large central common room with the throne and dining
tables, with the most commonly used things most immediately to hand by
the entrance (armoury, beds) and the more niche things (mine, kitchen,
storage) tucked away on the edge.

Spent a bit of time rearranging the goblin encampment. A big thing is
changing the overall layout, rather than being tucked away behind a
door, the throne-room and dining hall are one big central chamber that
all the other rooms snake off of. This makes it more of a common meeting
place because you'll be going through there to get to the other stuff,
rather than just going through the plain corridor between them.

The storage room and mine are niche rooms and are moved further away
from the entrance onto the western side. It has a little rat surprise
for dilligent miners. They're still accessible, but you aren't reminded
of how empty they are every time you leave the base. The kitchen is
tucked away closer to the dining hall, and connected directly to the
"outside" area (which is exclusively used for farming from what I
understand). I made the kitchen messier and made the 'outside' a bit
less depressing.

The chieftain's room is behind the chieftain's throne, and the
chieftain's hoard is behind the chieftain's room. I think it just makes
sense. Means the chieftain doesn't have to waste so much time going back
and forth, and it reflects a bit of personality, the chieftain is
self-centered and greedy, and keeps their treasures close. Mine mine
mine!.
To reflect a recent buff giving whip skill to goblin chieftain, the
chieftain also starts with a whip in their drawer. Since goblins can
definitely speak imperial now, there's no serious need to keep
talkstones around, taking a neck slot and having to throw one at your
interlocutor and hope they put it on and kindly give it back when you're
done talking. So instead, Cursed Collars! A much more fitting bit of
neck-based magical equipment for goblins to have.

The Shaman's room is also accessible from the little throne entrance and
from the prison area. The prison area has (another) rat-based surprise,
which can be eaten or kept as a pet or used to feed unlucky prisoners
to. It also has a NEW room, a combination surgery-and-torture-chamber
with some basic stone improvised surgical tools.

Lastly, instead of a steel dagger and three hunting knives, they start
with two iron daggers and two hunting knives. If you want steel, you'd
better steal it from someone. There's also a spare axe for goblin
Raiders and spare bows and arrows for rabble/boltslingers. Also,
importantly, they have a couple spare Maces for the brutes and for the
chieftain if he wants one.

Now, instead of having that really awkward room with a wandering saiga
in it that continuously blocks your way into the armoury, I was instead
going to include a little Stable room - but I thought it'd be funnier,
in that goblin way, to just have it kinda be chilling in the central
chamber. Up to the player goblins whether they kill and eat it or keep
it as a pet or for war. And with more space to roam, it's less likely to
actually be a problem.

Lastly, the area is generally more well decorated, with messes and
splats and stains and rats and poo littering the floor to make it feel
more lived in and goblin-y

## Why It's Good For The Game

The original encampment was drab and lonely and awkwardly laid out.
Goblins never went to the throne room because it was so out-of-the-way
and inconvenient, and making a quick stop to eat, sleep and restock on
gear had a lot of added tedium by the long, snaking corridor separating
everything. It should be a more fun area for a full warband to base
their operation in and get rowdy and drink together, and more convenient
and less lonely for solo goblins to use as a den between mischief.


## Proof of Testing (Required)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->


![image](https://github.com/user-attachments/assets/70394a07-e2f6-4349-8e26-70e29c746a06)

![image](https://github.com/user-attachments/assets/2903ac89-a434-463c-9581-6b38f400cb63)

![image](https://github.com/user-attachments/assets/2260a2ca-2eea-4cbd-8016-5f06cbc6fd80)

![image](https://github.com/user-attachments/assets/04a24541-1080-462b-86b7-1f30bbc14dde)

![image](https://github.com/user-attachments/assets/2b030729-85de-4cf5-953d-5398c5bd144d)

![image](https://github.com/user-attachments/assets/07679a47-eede-4ca4-ab15-677c37ee527e)