Added ability to recolour character chatlog text on a per-recipient basis (#2921)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR adds the ability to define chat colours to be perceived by a
given player when a different given player speaks. To offer an example,
if person A feeds person B a love potion, person B's text in the chat
box will appear in pink to indicate the supernatural influence they have
over person A.

This is achieved through a new list stored in each Mind of
"special_people". This list takes player character names as keys and
strings representing colour hash codes as values; when the owning Mind
receives a message, if the player is on their list, the associated
colour will be used when printing their text to the chat log.

This ability is currently applied in four scenarios:
- When a person is fed a love potion by another person (pink);
- When a person is a Vampyre Spawn, seeing the text of the Vampyre Lord
(red);
- When a person is a Death Knight, seeing the text of the Vampyre Lord
(red);
- When a person is a raised undead, seeing the text of the necromancer
or lich that raised them (purple).

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
At present, there is some amount of communication of the level of
influence the dominant party has in these relationships - typically in
the form of objective text or a chat message received at the moment that
the influence is first applied. However, this influence isn't really
communicated on an ongoing basis, which I think can lead to some - not
all - players underestimating it. I think applying special text colours
in cases where some manner of mental influence is especially profound,
with an appropriate colour to the type of influence, could go some way
towards remedying that situation.

There's an additional practical layer to this - in cases where, say, the
vampyre lord or lich has become powerful and has a lot of minions, there
might be a lot of chatter, making it hard to pick out the leader's
comments; this feature will make it _much_ easier for Vampyres to hear
the vampyre lord (at least in person - not, for the time being,
telepathically), and likewise for skeletons hearing their masters.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Initial testing showing the colour working alongside various types of
in-game formatting, including markdown, whispers, shouting, yelling, and
combined emotes and dialogue. Note that green is currently unused by any
implementation of this functionality - this was added via variable
editing for testing purposes.

![image](https://github.com/user-attachments/assets/dd1b5e31-f49f-4d92-b842-1da6d85abcec)

Image showing a friend, Skyglazer, receiving pink text after being fed a
love potion. (I can confirm I was seeing plain-coloured, rather than
pink, text). Pardon the contents of the actual text here; I'd previously
missed the hash from the code in the previous round of testing.

![image](https://github.com/user-attachments/assets/f2b70f2d-dd3e-42d1-a82e-072cb7ff78aa)

Image showing Skyglazer receiving red text as a Vampyre Spawn, when
speaking with the Vampyre Lord.

![image](https://github.com/user-attachments/assets/0818419e-c569-4856-a519-f548fc1ebe6c)

Image showing Skyglazer receiving purple text as a skeleton, having been
raised by the Lich. (Again, pardon the text; we determined an ideal
colour as we were testing, but the initial colour was successfully
applied.)

![image](https://github.com/user-attachments/assets/6375dbd9-2101-4d07-a389-3c7437de1d47)

And Skyglazer confirming, after dying, returning to the lobby, and being
given control of a new body, that these text colours were no longer
being applied:

![image](https://github.com/user-attachments/assets/6dcd2187-3dbf-4cef-b329-c605795aaa1e)

We were also able to confirm that the colours persist through a lich
dying and resurrecting, due to using the same real_name.

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->