Removal of Grenzel Kommandant (#2949)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR Flat out removes Grenzel Kommandant
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Grenzel's are a professional mercenary group yes. But at present,
Kommandant is a flavored role with stupid stats and stupid gear. Further
more, it's existance actively harms players that don't want to
participate in clique-y behavior, and at worst is used by players as a
gamer role without giving a damn about their supposed leadership needs.

All of that aside- Individual mercs should never have been given a
'leadership' role, beyond one for a general guild leadership. We do not
want to encourage large amounts of mercs to take the same job if
possible.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
N/A
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->