Cleans up auto mining code. (#2957)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

- Removes spawn (no idea what it was doing so I took it out)
- makes the while look dependent on density and user being adjacent
- breaks while if density is no longer true.

If this still seems wrong or could be somewhere better let me know.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Fixes a few edge cases with the previous code.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/181c2214-a161-4064-a856-1b60590a4dcd


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->