Removes refugee noble gender-based class selection. (#2917)
## About The Pull Request

This PR makes it so the Refugee Noble's subclass selection is the same
regardless of character gender, renaming the existing subclasses into
gender-neutral ones and making each of them available for both genders.
The "widowed consort" role even gives male characters a shirt and pants
rather than a dress.

## Why It's Good For The Game

The subclasses being gender-locked strikes me as more of a holdover from
older versions of RT lore where women weren't allowed in combat roles.
In our server, women can be knights, mercenaries, templars, all sorts of
warriors-- so it makes sense to allow them to pick the archetypes
available to male characters in the Noble refugee role.

## Proof of Testing (Required)

Forgot to use the test map for this one, but here is a video of me
spawning in as each role with a character of both genders.
https://www.youtube.com/watch?v=EnSS5QFic6g