adds fingering... (#2932)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
See title
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
community request (2 people). It's virtually the same as the other but
I'll let whoever has to approve this and who wants it in argue it out, I
just did it on their behalf.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Self:


![image](https://github.com/user-attachments/assets/df2a235c-4d1c-40d8-9ae4-ff87cd701b48)

![image](https://github.com/user-attachments/assets/e71d5703-f4c0-4c9e-bfe9-dba961db0845)
Other:


![image](https://github.com/user-attachments/assets/184ebea0-a98f-4690-818d-6baa79d07a63)

![image](https://github.com/user-attachments/assets/dc405826-aec4-4cf1-b57b-18411c75e7e5)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->